Moved shortcuts.py
