+++ corehq/apps/linked_domain/updates.py
@@ -5,8 +5,6 @@ from functools import partial
 from django.utils.translation import ugettext as _
 from django.db import transaction
 
-from toggle.shortcuts import set_toggle
-
 from corehq.apps.data_interfaces.models import (
     AutomaticUpdateRule, CaseRuleAction, CaseRuleCriteria,
     ClosedParentDefinition, CustomActionDefinition,
@@ -101,6 +99,7 @@ from corehq.apps.linked_domain.ucr import update_linked_ucr
 from corehq.apps.linked_domain.keywords import update_keyword
 from corehq.apps.locations.views import LocationFieldsView
 from corehq.apps.products.views import ProductFieldsView
+from corehq.apps.toggle_ui.shortcuts import set_toggle
 from corehq.apps.userreports.dbaccessors import get_report_configs_for_domain
 from corehq.apps.userreports.util import (
     get_static_report_mapping,
