+++ corehq/apps/toggle_ui/shortcuts.py
