+++ corehq/apps/toggle_ui/views.py
@@ -16,6 +16,7 @@ from corehq.apps.domain.decorators import require_superuser_or_contractor
 from corehq.apps.hqwebapp.decorators import use_datatables
 from corehq.apps.hqwebapp.views import BasePageView
 from corehq.apps.toggle_ui.models import ToggleAudit
+from corehq.apps.toggle_ui.shortcuts import parse_toggle, namespaced_item
 from corehq.apps.toggle_ui.tasks import generate_toggle_csv_download
 from corehq.apps.toggle_ui.utils import find_static_toggle
 from corehq.apps.users.models import CouchUser
@@ -39,7 +40,6 @@ from corehq.util.soft_assert import soft_assert
 from couchforms.analytics import get_last_form_submission_received
 from soil import DownloadBase
 from toggle.models import Toggle
-from toggle.shortcuts import parse_toggle, namespaced_item
 
 NOT_FOUND = "Not Found"
